Make the main loop match watchdog's quickstart more closely
Everything in the quickstart's loop is there for a reason.

Most notably, the use of a `finally` block rather than a bare `except`
block. Bare `except` handles every exception, including things you
*really* don't want to catch like out-of-memory conditions. The
quickstart wants to _finalize_ the observer-stop its thread and reclaim
those resources-when exiting the scope it's supposed to be running, but
if the scope is being exited for an abnormal reason we shouldn't
continue on as though everything is fine. `finally` gives us those
semantics, while `except:` empathically does not.

Secondly, the observer thread can terminate for reasons other than the
main thread requesting it stop. Should this occur the original code
would loop infinitely; this new version will instead promptly exit.
Detecting and reporting such a condition is left as an exercise for the
reader.

As a side note, the printout that the observer stopped is moved to after
the thread join. In its original position it did not accurately reflect
the state of the world: the observer had only been _requested_ to stop.
We can only say that it _has_ stopped once we've observed control return
from the thread join.