Eschew global variables, reuse network clients
It's essentially universally accepted that global variables are not a
good practice, so here we pass them down to where they are needed as
parameters. In order to illustrate a motivation for the change, we
extend the capabilities of `OnMyWatch` slightly, allowing any number of
independent handlers to be registered.

While we're in there adding parameters, we hoist the creation of the S3
client from deep in the guts of the Handler class to the top level. This
is exactly the "inversion of control" or "dependency injection" pattern
familiar from other languages and gives the same benefits: since client
code determines the runtime type of the client, it is trivial to pass a
mock implementation for testing, and so on. In addition, reusing the
client allows the SDK to transparently reuse HTTPS connections as
appropriate, saving connection overhead.