Remove repetition in event handler
The body of every conditional branch in the event handler was
essentially identical, so we can collapse the differences into the
contidional itself.

There's also a meaningful semantic change here, which I will insist is
correct based on my experience with file watchers: instead of firing on
modifications, this version fires when the file which was opened for
modifications is closed. In my experience, watchers that fire on every
call to `write(2)` are poorly behaved, because there are usually _many_
such calls involved in real-world edits and you end up in a race with
the program making the edits and it's really just not a good time.
_Also,_ at least on Linux with the inotify backend, a program which
chooses to `mmap` the file to modify it will not generate the modified
events at all. Waiting for the file to be closed does not _eliminate_
all race conditions, but it does minimize common ones.

Finally, I will simply note that this version retains the original
semantic bug regarding matching the file: because it uses a simple
substring match, a request to watch a file `foo` will also match
modifications to `foo_bar`. How this is best corrected depends on the
exact semantics of the watchdog package's interface, and I don't know
them.